requirements.txt
notion-client>=2.2.0
python-telegram-bot>=20.7
anthropic>=0.25.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
reportlab>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
